+++ packages/unity-js-bridge/src/thirdweb-bridge.ts
@@ -144,7 +144,7 @@ class ThirdwebBridge implements TWBridge {
       }
       (globalThis as any).X_SDK_NAME = "UnitySDK_WebGL";
       (globalThis as any).X_SDK_PLATFORM = "unity";
-      (globalThis as any).X_SDK_VERSION = "4.7.0";
+      (globalThis as any).X_SDK_VERSION = "4.7.5";
       (globalThis as any).X_SDK_OS = browser?.os ?? "unknown";
     }
     this.initializedChain = chain;
@@ -244,6 +244,9 @@ class ThirdwebBridge implements TWBridge {
             paymasterUrl: sdkOptions.smartWalletConfig?.paymasterUrl,
             // paymasterAPI: sdkOptions.smartWalletConfig?.paymasterAPI,
             entryPointAddress: sdkOptions.smartWalletConfig?.entryPointAddress,
+            erc20PaymasterAddress:
+              sdkOptions.smartWalletConfig?.erc20PaymasterAddress,
+            erc20TokenAddress: sdkOptions.smartWalletConfig?.erc20TokenAddress,
           };
           walletInstance = new SmartWallet(config);
           break;
@@ -381,6 +384,7 @@ class ThirdwebBridge implements TWBridge {
           personalWallet,
           authOptions,
         );
+        await this.switchNetwork(chainId); // workaround for polygon/mumbai
         if (this.activeWallet) {
           // Pass EOA and reconnect to initialize smart wallet
           await this.initializeSmartWallet(
@@ -513,7 +517,10 @@ class ThirdwebBridge implements TWBridge {
           // ccipReadEnabled: txInput.ccipReadEnabled,
         });
 
-        if (routeArgs[2].includes("send")) {
+        if (routeArgs[2].includes("sign")) {
+          const result = await tx.sign();
+          return JSON.stringify({ result: result }, bigNumberReplacer);
+        } else if (routeArgs[2].includes("send")) {
           tx.setGaslessOptions(
             routeArgs[2] === "sendGasless"
               ? this.activeSDK.options.gasless