+++ packages/wallets/test/smart-wallet-integration.test.ts
@@ -12,7 +12,8 @@ let smartWallet: SmartWallet;
 let smartWalletAddress: string;
 let personalWallet: LocalWallet;
 let contract: SmartContract;
-const factoryAddress = "0x13947435c2fe6BE51ED82F6f59C38617a323dB9B";
+const factoryAddress = "0x13947435c2fe6BE51ED82F6f59C38617a323dB9B"; // pre 712
+const factoryAddressV2 = "0xC64d04AedecA895b3F20DC6866b4b532e0b22634"; // post 712
 const chain = Mumbai;
 
 const describeIf = (condition: boolean) =>
@@ -98,7 +99,33 @@ describeIf(!!process.env.TW_SECRET_KEY)("SmartWallet core tests", () => {
     expect(balance.toNumber()).toEqual(7);
   });
 
-  it("can sign and verify 1271", async () => {
+  it("can sign and verify 1271 old factory", async () => {
+    const message = "0x1234";
+    const sig = await smartWallet.signMessage(message);
+    const isValidV1 = await smartWallet.verifySignature(
+      message,
+      sig,
+      smartWalletAddress,
+      chain.chainId,
+    );
+    expect(isValidV1).toEqual(true);
+    const isValidV2 = await checkContractWalletSignature(
+      message,
+      sig,
+      smartWalletAddress,
+      chain.chainId,
+    );
+    expect(isValidV2).toEqual(true);
+  });
+
+  it("can sign and verify 1271 new factory", async () => {
+    smartWallet = new SmartWallet({
+      chain,
+      factoryAddress: factoryAddressV2,
+      gasless: true,
+      secretKey: process.env.TW_SECRET_KEY,
+    });
+    smartWalletAddress = await smartWallet.connect({ personalWallet });
     const message = "0x1234";
     const sig = await smartWallet.signMessage(message);
     const isValidV1 = await smartWallet.verifySignature(